requirements.txt
chromadb==1.5.1
groq==0.37.1
langchain-community==0.4.1
langchain-groq==1.1.2
pandas==3.0.1
python-dotenv==1.2.1
ragas==0.4.3
sentence-transformers==5.2.3
xlrd==2.0.2